feat(bm/mr/gp): Server Component shells pre-fetch initial qualification data (#693)
- Create `src/lib/api-factories/qual-initial-data.ts`:
  `fetchQualInitialData(config, id)` runs the same Prisma queries as
  GET /api/tournaments/[id]/{bm,mr,gp} (qualifications + matches +
  computeQualificationRanks + players) in a single parallel batch;
  returns null on error so the client falls back gracefully.
- Rename each mode page.tsx → page-client.tsx; component now accepts
  `tournamentId: string` and `initialData` props instead of `use(params)`.
  Add `initialData` to each usePolling call.
- New page.tsx for each mode: `force-dynamic` Server Components that
  await params, call `fetchQualInitialData`, and render the client shell.

Together with the TA migration (previous commit), all four mode pages
now pre-fetch initial data server-side, fulfilling issue #693.
